auth_service/requirements.txt
Django==5.2.3
djangorestframework
djoser
djangorestframework-simplejwt
psycopg2-binary